obstruction function working, and formatting + naming changed.

+++ path-pilot.X/sensors.h
@@ -8,13 +8,15 @@
 #ifndef SONIC_H
 #define	SONIC_H
 
+#include <avr/io.h>
+
 typedef enum {
     Red,
     Green,
     None
 } color_t;
 
-double get_distance();
+uint16_t get_distance();
 
 void set_led(color_t color);
